refactor: rename utils/create_node_helper.py to utils/detect_id_mapper.py
