+++ peekingduck/commands/create_node_helper.py
@@ -24,9 +24,6 @@ import click
 
 from peekingduck.commands import LOGGER_NAME
 
-# Master map file for class name to object IDs for object detection models
-MASTER_MAP = "pipeline/nodes/model/master_map.yml"
-
 logger = logging.getLogger(LOGGER_NAME)  # pylint: disable=invalid-name
