+++ peekingduck/pipeline/nodes/abstract_node.py
@@ -23,7 +23,7 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
 from peekingduck.config_loader import ConfigLoader
-from peekingduck.utils.create_node_helper import obj_det_change_class_name_to_id
+from peekingduck.utils.detect_id_mapper import obj_det_change_class_name_to_id
 
 
 class AbstractNode(metaclass=ABCMeta):
