+++ tests/pipeline/nodes/test_node.py
@@ -17,7 +17,7 @@ from typing import Dict
 import pytest
 
 from peekingduck.pipeline.nodes.abstract_node import AbstractNode
-from peekingduck.utils.create_node_helper import obj_det_change_class_name_to_id
+from peekingduck.utils.detect_id_mapper import obj_det_change_class_name_to_id
 
 
 class ConcreteNode(AbstractNode):
